third_party/WebKit/LayoutTests/inspector/console/console-message-from-inline-with-url.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>
function foo()
{
    console.log("foo");
}
function addInlineWithSyntaxError()
{
    var headElement = document.getElementsByTagName("head")[0];
    var scriptElement = document.createElement("script");
    scriptElement.setAttribute("type", "text/javascript");
    headElement.appendChild(scriptElement);
    scriptElement.text = "}\n//# sourceURL=boo.js";
}
//# sourceURL=foo.js
</script>
<script>
function test() {
  TestRunner.evaluateInPage('setTimeout(foo, 0);', ConsoleTestRunner.waitUntilMessageReceived.bind(this, step2));

  function step2() {
    TestRunner.evaluateInPage(
        'setTimeout(addInlineWithSyntaxError, 0);', ConsoleTestRunner.waitUntilMessageReceived.bind(this, step3));
  }

  function step3() {
    ConsoleTestRunner.dumpConsoleMessages();
    TestRunner.completeTest();
  }
}
</script>
</head>
<body onload="runTest()">
<p>
Tests that log message and syntax errors from inline scripts with sourceURL are logged into console, contains correct link and doesn't cause browser crash.
</p>
</body>
</html>
